minor: Disable the use of /scuba if you are not on duty/or on a different character

+++ Normal Build/core_modules2/diving/client.lua
@@ -170,10 +170,12 @@ end)
 
 RegisterNetEvent("diving:emergency")
 AddEventHandler("diving:emergency", function()
-	if ScubaHas() then
-		Scuba.delete()
-		Scuba.reset(PlayerPedId())
-	else
-		Scuba.equip(PlayerPedId())
+	if exports.policejob:getIsInService() then
+		if ScubaHas() then
+			Scuba.delete()
+			Scuba.reset(PlayerPedId())
+		else
+			Scuba.equip(PlayerPedId())
+		end
 	end
 end)
